requirements.txt
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.12.0
APScheduler==3.10.4
arrow==1.4.0
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
boto3==1.35.0
blessed==1.25.0
celery==5.6.0
Cerberus==1.3.8
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==2.3.0
Django==6.0.1
django-allauth==0.57.0
django-axes==8.1.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.2
django-picklefield==3.4.0
django-q==1.3.9
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
jinxed==1.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
multidict==6.7.0
numpy==2.4.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
postgrest==2.25.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pypng==0.20220715.0
pytesseract==0.3.13
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
qrcode==7.4.2
realtime==2.25.0
redis==3.5.3
referencing==0.37.0
reportlab==4.4.6
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
sentry-sdk==2.48.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.7.0
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
gunicorn